# openapi-v2.yaml
openapi: 3.1.0
info:
  title: GPT Paywall API - Enhanced Version
  version: 2.0.0
  description: Enhanced API for managing GPT paywall access, payments, and user management
servers:
  - url: https://your-app.vercel.app
    description: Production server on Vercel
components:
  securitySchemes:
    gptApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for GPT authentication
  schemas:
    AccessResponse:
      type: object
      properties:
        has_access: 
          type: boolean
          description: Whether the user has premium access
        plan: 
          type: string
          nullable: true
          description: Current plan name
        status: 
          type: string
          nullable: true
          description: Subscription status
        current_period_end: 
          type: string
          nullable: true
          format: date-time
          description: When the current billing period ends
    CheckoutRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User email address
        plan:
          type: string
          default: "pro"
          description: Subscription plan
    CheckoutResponse:
      type: object
      properties:
        session_id:
          type: string
          description: Stripe checkout session ID
        checkout_url:
          type: string
          description: Direct URL to Stripe checkout
    PortalRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User email address
    PortalResponse:
      type: object
      properties:
        portal_url:
          type: string
          description: URL to Stripe customer portal
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        plan:
          type: string
        status:
          type: string
        customer_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            hasMore:
              type: boolean
    AnalyticsResponse:
      type: object
      properties:
        total_users:
          type: integer
        active_subscriptions:
          type: integer
        pending_users:
          type: integer
        cancelled_users:
          type: integer
        plans:
          type: object
        revenue_estimate:
          type: number
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
paths:
  /api:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  timestamp:
                    type: string
  /api/check-access:
    get:
      summary: Check if a user has premium access
      operationId: checkAccess
      security:
        - gptApiKey: []
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
          description: User email address
      responses:
        "200":
          description: Access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/user/{email}:
    get:
      summary: Get detailed user information
      operationId: getUser
      security:
        - gptApiKey: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
            format: email
          description: User email address
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
  /api/stripe/create-checkout-session:
    post:
      summary: Create a Stripe checkout session
      operationId: createCheckoutSession
      security:
        - gptApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /api/stripe/create-portal-session:
    post:
      summary: Create a Stripe customer portal session
      operationId: createPortalSession
      security:
        - gptApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalRequest"
      responses:
        "200":
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResponse"
  /api/airtable/users:
    get:
      summary: List all users with pagination and filters
      operationId: listUsers
      security:
        - gptApiKey: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: status
          schema:
            type: string
            enum: [active, pending, cancelled]
        - in: query
          name: plan
          schema:
            type: string
            enum: [free, pro, premium, enterprise]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
    post:
      summary: Create or update a user
      operationId: createOrUpdateUser
      security:
        - gptApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                plan:
                  type: string
                status:
                  type: string
                customFields:
                  type: object
      responses:
        "200":
          description: User created or updated
  /api/airtable/users/{email}:
    delete:
      summary: Delete a user
      operationId: deleteUser
      security:
        - gptApiKey: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: User deleted successfully
        "404":
          description: User not found
  /api/airtable/analytics:
    get:
      summary: Get analytics and statistics
      operationId: getAnalytics
      security:
        - gptApiKey: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
  /api/airtable/users/bulk:
    post:
      summary: Perform bulk operations on users
      operationId: bulkUserOperations
      security:
        - gptApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - emails
              properties:
                action:
                  type: string
                  enum: [update, delete]
                emails:
                  type: array
                  items:
                    type: string
                    format: email
                updates:
                  type: object
                  description: Fields to update (required for update action)
      responses:
        "200":
          description: Bulk operation results